обновление сессии

+++ src/main/java/com/sergeybochkov/transmissionremote/model/Size.java
@@ -1,20 +1,18 @@
 package com.sergeybochkov.transmissionremote.model;
 
-import java.math.BigDecimal;
-
 public final class Size {
 
     private static final String[] SIZES = {"B", "kB", "MB", "GB", "TB"};
 
-    private final BigDecimal size;
+    private final double size;
 
-    public Size(BigDecimal size) {
+    public Size(double size) {
         this.size = size;
     }
 
     @Override
     public String toString() {
-        double res = size.doubleValue();
+        double res = size;
         int count = 0;
         while (res >= 1000) {
             res /= 1000;
